fix: Change pips from int to double for correct price formatting
- Changed pips field from int to double in SignalData and OrderFlowStrategyEnhanced
- Fixed IllegalFormatConversionException when formatting prices with %f
- Risk calculations now use raw tick offsets (not multiplied by pips)
- Updated VWAPCalculator and EMACalculator to accept double pips
- Fixed test cases for tick offset calculations

The root cause was that pips (pip size, e.g., 0.25 for ES) was stored as int,
truncating to 0. When int * int produces int, the %f format specifier fails.

+++ Strategies/src/main/java/velox/api/layer1/simplified/demo/EMACalculator.java
@@ -60,7 +60,7 @@ public class EMACalculator {
     /**
      * Get distance in ticks from EMA
      */
-    public double getDistance(double price, int pips) {
+    public double getDistance(double price, double pips) {
         if (Double.isNaN(ema)) return 0;
         return (price - ema) / pips;
     }
